Wrote out a function to print out the tune log values. Also manually added it to the NAMESPACE (due to not having roxygen2) and updated rawrr.R to call the function.
Added command that reads Sequence files (.sld) and extracts all available information.

Added command to extract Liquid Chromatopgraphy (LC) pump pressures over time.
This assumes that the pressure values are stored as an A/D Card Detector Trace, as opposed to a Status Log associated with the A/D Card instrument.